Исправление бага при экипировке пистолета и меча/топора.
